# src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
            "Synonym": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "from": {
                        "type": "string"
                    },
                    "target": {
                        "type": "string"
                    },
                    "filenameOverride": {
                        "type": "string"
                    },
                    "exemptions": {
                        "type": "string"
                    }
                },
                "required": ["id", "from", "target", "exemptions"],
                "type": "object",
                "additionalProperties": true
            },
            "SearchHistoryEntry": {
                "properties": {
                    "term": {
                        "type": "string"
                    },
                    "results": {
                        "type": "number",
                        "format": "double"
                    },
                    "appId": {
                        "type": "string"
                    },
                    "series": {
                        "type": "number",
                        "format": "double"
                    },
                    "episode": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": ["term", "results"],
                "type": "object",
                "additionalProperties": true
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
    "info": {
        "title": "iplayarr",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "paths": {
        "/api/synonyms": {
            "get": {
                "operationId": "GetAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Synonym"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["Synonyms"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "Create",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Synonym"
                                }
                            }
                        }
                    }
                },
                "tags": ["Synonyms"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Synonym"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Update",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Synonym"
                                }
                            }
                        }
                    }
                },
                "tags": ["Synonyms"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Synonym"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "Delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["Synonyms"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/synonyms/searchHistory": {
            "get": {
                "operationId": "SearchHistory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SearchHistoryEntry"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["Synonyms"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/synonyms/lookup/{appId}": {
            "get": {
                "operationId": "Lookup",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SearchHistoryEntry"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["Synonyms"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "appId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "term",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}